backend/pyproject.toml
[project]
name = "py-firefly-iii-tools"
version = "0.4.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.121.2",
    "fireflyiii-enricher-core",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv==1.2.1",
    "python-multipart>=0.0.20",
    "streamlit==1.51.0",
    "tabulate>=0.9.0",
    "uvicorn>=0.38.0",
]

[tool.uv.sources]
fireflyiii-enricher-core = { git = "https://github.com/wini83/fireflyiii-enricher-core.git", rev = "main" }


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
[dependency-groups]
dev = [
    "black>=25.11.0",
    "commitizen>=4.10.0",
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "ruff>=0.14.5",
]
